automated-genome-qc
This repository contains an uncompressed and executable version of a bioinformatics pipeline designed to perform quality control on genomic data. It includes automated steps for read quality assessment, adapter trimming, and basic statistics reporting. Suitable for whole genome or exome sequencing data.

Limitations: It is designed for a specific set of files, will be updated soon